// src/api/generated/models/PartidoDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PartidoDTO = {
    id: number;
    equipoLocal: string;
    equipoVisitante: string;
    fechaHora: string;
    grupo?: string;
    grupoId?: number;
    equipoLocalId?: number;
    equipoVisitanteId?: number;
    golesLocalReal?: number;
    golesVisitanteReal?: number;
    estado: string;
    minutosParaInicio?: number;
    minutosJugados?: number;
};